+++ PlayerReturn.cpp
@@ -25,7 +25,7 @@
 **************************************/
 void PlayerReturn::OnUpdate(Player *entity)
 {
-	float t = entity->cntFrame / 120;
+	float t = (float)entity->cntFrame / 120;
 	entity->pos = Easing<D3DXVECTOR3>::GetEasingValue(t, &entity->initpos, &PLAYER_INIT_POS, EasingType::OutExponential);
 
 };
